Attempt to add "text string" functionality.
Test added to testsuite.js. Note the unwanted multiply symbol after 'Vector:'.
Need some way of specifying whether the text expression should really just be a label,
as in this case, or should actually multiply, as in the 3x*'length' case, at the bottom of the fraction.
Also, probably need to display the multiply symbol if a text expression is multiplied.